economy/ledger: fix size_t/uint64 decode arg + struct-padding memcmp in test
Two correctness fixes on the accounting-core PR before merge:

- ledger.c: ants_ledger_record_decode declared n_pairs as uint64_t and
  passed &n_pairs to ants_cbor_dec_map, whose parameter is size_t *. On
  LP64 Linux size_t == unsigned long == 64-bit so it compiled clean, but
  on macOS size_t is unsigned long and uint64_t is unsigned long long —
  a -Wincompatible-pointer-types warning and a genuine type-punning bug.
  Use size_t, matching the codec signature and the inference/orchestration
  precedent.

- test_ledger.c: test_record_round_trip_zeroes compared decoded vs
  original with memcmp over the whole struct, which reads the struct's
  trailing padding bytes (UB, and they differ between a memset-init'd and
  a field-by-field-decoded record). Compare the seven logical fields
  individually instead.

Full ctest 15/15 green; -Wall -Wextra -Wpedantic + clang-format clean,
0 warnings on a full ants_ledger rebuild.

+++ economy/ledger/src/ledger.c
@@ -304,7 +304,7 @@ static ants_error_t decode_bool_field(ants_cbor_dec_t *dec, uint64_t key, bool *
 ants_error_t ants_ledger_record_decode(const uint8_t *buf, size_t len, ants_ledger_peer_t *rec)
 {
     ants_cbor_dec_t dec;
-    uint64_t n_pairs = 0;
+    size_t n_pairs = 0;
     uint64_t u;
     ants_error_t rc;
